try to make sure that we can run the scripts without horovod
